requirements.txt
openai>=1.40.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pysbd>=0.3.4
streamlit>=1.36.0
pandas>=2.2.0
altair>=5.0.0
